requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
torch
open_clip_torch